Make Drush an optional dependency. Improve output.
Let Drupid run even if Drush is not found. Since
Drush is currently used only to check whether
Drupal is bootstrapped or whether a project is
installed, and such tests are invoked only when
downgrading a project, Drupid may simply log an
error after a downgrading action if Drush is not
installed in the system. As usual, --force allows
the user to override Drupid's behavior.

Use warnings instead of logging errors when
updating with --force. Also make sure that warnings
and errors are printed *after* logging the
corresponding action.